Add ImportedPiece.regularize method
This is preliminary support for returning observations at regular
offsets. Depending on the type of result, this will work or not. For
those that don't work yet, they will be provided with a `unit`
parameter like this commit does for the getMelodic method and this will
provide the same functionality for these cases.