Add combineRests combineUnisons settings
These settings are added to .getNoteRest. `combineRests` defaults to
true which slightly changes a core behavior. When True, it removes
non-first consecutive rests. Similarly, combineUnisons removes non-first
consecutive attacks of the same note in the same voice. However,
`combineRests` defaults to False because this is removing actual
attacks from the representation of the piece.