Refine cadence and cvf detection
Now we differentiate between the two types of bass evasion with
different labels. We also add a vetting step for evaded bassizans "b"
labels when they have no preparation as these are usually actually
chanson idioms. Also disregard any Tenorizans or evaded Tenorizans cvfs
if a Bassizans, "B", is detected, because the Tenorizans won't play any
role in the cadence classification in this case, so ignoring them
greatly reduces the number of labels we need in cadenceLabels.csv.