Update cvf and cad csv files to use 1
We were previously using unisons. This also removes about 6 patterns
from the cvf csv file because they are duplicates now that we
collapsed "Held" and "1" to just "1". Also, there was an issue with
one of the "Q" and "A" patterns in that, on its own, it was
indistinguishable from a "B" "C" pattern. So now "A" can't get
overwritten by "C", and if a "C" tries to, its accompanying "B"
gets overwritten to "Q".