Use regex to match cadenceLabels.csv
This is a little more complicated, but it makes it possible to add
things like z? to allow for abandoned Tenorizans cvfs, for example,
without changing the analysis or bloating the file. It also adds a
"Pattern" which is the old "Keys" column and "Keys" has become the
regex keys used for matching. Those columns are only seen in the
.cadences results when keep_keys=True.