Add .file_name, .modelFinder, and 'thematic' parameter
Big commit. This fixes several bugs and warnings. It adds the
ImportedPiece.fine_name attribute which is the file name without
the file extension. Not sure if this works properly on Windows.
A "thematic" setting was also added for the .entries method to
filter out entries that only happen once. The biggest addition is
of CorpusBase.modelFinder. This could be further refined but does
work as is.